delete findstation and add my car in slidebar
