+++ src/components/driver/DriverProfile.tsx
@@ -17,21 +17,11 @@ import { useLanguage } from "../LanguageContext";
 import { User as UserType } from "../../App";
 import { Battery } from "lucide-react";
 
-interface Vehicle {
-  id: string;
-  make: string;
-  model: string;
-  year: number;
-  vin: string;
-  batteryModel: string;
-}
-
 interface DriverProfileProps {
   user: UserType;
   profileName: string;
   profileEmail: string;
   profilePhone: string;
-  vehicles: Vehicle[];
   onNameChange: (name: string) => void;
   onEmailChange: (email: string) => void;
   onPhoneChange: (phone: string) => void;
@@ -42,7 +32,6 @@ export function DriverProfile({
   profileName,
   profileEmail,
   profilePhone,
-  vehicles,
   onNameChange,
   onEmailChange,
   onPhoneChange,
@@ -122,75 +111,7 @@ export function DriverProfile({
         </CardContent>
       </Card>
 
-      {/* Vehicle Information */}
-      <Card className="border border-orange-500 rounded-lg">
-        <CardHeader>
-          <CardTitle className="flex items-center space-x-2 text-orange-500 fond-bold">
-            <Car className="w-5 h-5" />
-            <span>{t("driver.vehicleInformation")}</span>
-          </CardTitle>
-          <CardDescription>
-            {t("driver.vehicleInformationDesc")}
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <div className="space-y-4">
-            {vehicles.map((vehicle) => (
-              <div
-                key={vehicle.id}
-                className="flex items-center justify-between p-4 border rounded-lg text-orange-500 fond-bold"
-              >
-                <div className="flex items-center space-x-3">
-                  <Car className="w-8 h-8 text-blue-500" />
-                  <div>
-                    <p className="font-medium">
-                      {vehicle.year} {vehicle.make} {vehicle.model}
-                    </p>
-                    <p className="text-sm text-gray-500">
-                      {t("driver.battery")} {vehicle.batteryModel}
-                    </p>
-                    <p className="text-xs text-gray-400">VIN: {vehicle.vin}</p>
-                  </div>
-                </div>
-                <div className="flex items-center space-x-2">
-                  <Badge className="bg-orange-500 text-white" variant="secondary">{t("driver.primary")}</Badge>
-                  <Button size="sm" variant="outline" className="bg-orange-500 text-white">
-                    <Edit className="w-4 h-4" />
-                  </Button>
-                </div>
-              </div>
-            ))}
-            <Button variant="outline" className="w-full bg-orange-500 text-white">
-              {t("driver.addVehicle")}
-            </Button>
-          </div>
-        </CardContent>
-      </Card>
 
-      {/* Driver Stats */}
-      <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-        <Card className="border border-orange-500 rounded-lg">
-          <CardContent className="p-4 text-center">
-            <Battery className="w-8 h-8 text-green-500 mx-auto mb-2" />
-            <p className="text-2xl font-bold">127</p>
-            <p className="text-sm text-gray-500">{t("driver.totalSwaps")}</p>
-          </CardContent>
-        </Card>
-        <Card className="border border-orange-500 rounded-lg">
-          <CardContent className="p-4 text-center">
-            <Star className="w-8 h-8 text-yellow-500 mx-auto mb-2" />
-            <p className="text-2xl font-bold">4.9</p>
-            <p className="text-sm text-gray-500">{t("driver.averageRating")}</p>
-          </CardContent>
-        </Card>
-        <Card className="border border-orange-500 rounded-lg">
-          <CardContent className="p-4 text-center">
-            <Shield className="w-8 h-8 text-blue-500 mx-auto mb-2" />
-            <p className="text-2xl font-bold">$1,847</p>
-            <p className="text-sm text-gray-500">{t("driver.totalSavings")}</p>
-          </CardContent>
-        </Card>
-      </div>
 
       {/* Payment Methods */}
       <Card className="border border-orange-500 rounded-lg">
